% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemset.R
\name{itemset}
\alias{itemset}
\alias{itemset.default}
\title{Itemset Meta-features}
\usage{
itemset(...)

\method{itemset}{default}(x, features = "all", summary = c("mean", "sd"), ...)
}
\arguments{
\item{...}{Further arguments passed to the summarization functions.}

\item{x}{A data.frame contained only the input attributes.}

\item{features}{A list of features names or \code{"all"} to include all them.}

\item{summary}{A list of summarization functions or empty for all values. See
\link{post.processing} method to more information. (Default: 
\code{c("mean", "sd")})}
}
\value{
A list named by the requested meta-features.
}
\description{
Itemset characterization features measure measure the distribution of values 
of both single attributes and pairs of attributes.
}
\details{
The following features are allowed for this method:
 \describe{
   \item{"oneitemset"}{Individual frequency of each attributes' value.}
   \item{"twoitemset"}{Correlation information of the two attributes' 
     value pairs.}
 }
}
\examples{
## Extract all meta-features using formula
itemset(iris[,1:4])

## Extract some meta-features
itemset(iris[,1:4], c("oneitemset"))

## Use another summarization function
itemset(iris[,1:4], summary=c("min", "median", "max"))
}
\references{
Song, Q., Wang, G., & Wang, C. (2012). Automatic recommendation of 
  classification algorithms based on data set characteristics. Pattern 
  Recognition, 45(7), 2672-2689.
  
Wang, G., Song, Q., & Zhu, X. (2015). An improved data characterization 
  method and its application in classification algorithm recommendation. 
  Applied Intelligence, 43(4), 892-912.
}
\seealso{
Other meta-features: 
\code{\link{general}()},
\code{\link{infotheo}()},
\code{\link{statistical}()}
}
\concept{meta-features}
