% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical.R
\name{statistical}
\alias{statistical}
\alias{statistical.default}
\title{Statistical meta-features}
\usage{
statistical(...)

\method{statistical}{default}(
  x,
  features = "all",
  summary = c("mean", "sd"),
  transform = TRUE,
  ...
)
}
\arguments{
\item{...}{Further arguments passed to the summarization functions.}

\item{x}{A data.frame contained only the input attributes.}

\item{features}{A list of features names or \code{"all"} to include all them.
The details section describes the valid values for this group.}

\item{summary}{A list of summarization functions or empty for all values. See
\link{post.processing} method to more information. (Default: 
\code{c("mean", "sd")})}

\item{transform}{A logical value indicating if the categorical attributes
should be transformed. If \code{FALSE} they will be ignored. (Default: 
\code{TRUE})}
}
\value{
A list named by the requested meta-features.
}
\description{
Statistical meta-features are the standard statistical measures to describe
the numerical properties of a distribution of data. As it requires only
numerical attributes, the categorical data are transformed to numerical.
}
\details{
The following features are allowed for this method:
 \describe{
   \item{"cor"}{Absolute attributes correlation, which measure the 
   correlation between each pair of the numeric attributes in the dataset 
   (multi-valued). This measure accepts an extra argument called 
   \code{method = c("pearson", "kendall", "spearman")}. See 
   \code{\link[stats]{cor}} for more details.}
   \item{"cov"}{Absolute attributes covariance, which measure the covariance 
   between each pair of the numeric attributes in the dataset 
   (multi-valued).}
   \item{"eigenvalues"}{Eigenvalues of the covariance matrix (multi-valued).}
   \item{"gMean"}{Geometric mean of attributes (multi-valued).}
   \item{"hMean"}{Harmonic mean of attributes (multi-valued).}
   \item{"iqRange"}{Interquartile range of attributes (multi-valued).}
   \item{"kurtosis"}{Kurtosis of attributes (multi-valued).}
   \item{"mad"}{Median absolute deviation of attributes (multi-valued).}
   \item{"max"}{Maximum value of attributes (multi-valued).}
   \item{"mean"}{Mean value of attributes (multi-valued).}
   \item{"median"}{Median value of attributes (multi-valued).}
   \item{"min"}{Minimum value of attributes (multi-valued).}
   \item{"nrCorAttr"}{Number of attributes pairs with high correlation.}
   \item{"nrNorm"}{Number of attributes with normal distribution. The 
   Shapiro-Wilk Normality Test is used to assess if an attribute is or not is
   normally distributed.}
   \item{"nrOutliers"}{Number of attributes with outliers values. The 
   Turkey's boxplot algorithm is used to compute if an attributes has or does 
   not have outliers.}
   \item{"range"}{Range of Attributes (multi-valued).}
   \item{"sd"}{Standard deviation of the attributes (multi-valued).}
   \item{"skewness"}{Skewness of attributes (multi-valued).}
   \item{"tMean"}{Trimmed mean of attributes (multi-valued). It is the 
   arithmetic mean excluding the 20\% of the lowest and highest instances.}
   \item{"var"}{Attributes variance (multi-valued).}
 }
 This method uses simple binarization to transform the categorical attributes
 when \code{transform=TRUE}.
}
\examples{
## Extract all meta-features
statistical(iris[,1:4])

## Extract some meta-features
statistical(iris[1:4], c("cor", "nrNorm"))

## Extract all meta-features without summarize the results
statistical(iris[,1:4], summary=c())

## Use another summarization function
statistical(iris[,1:4], summary=c("min", "median", "max"))

## Do not transform the data (using only categorical attributes)
statistical(iris[,1:4], transform=FALSE)
}
\references{
Ciro Castiello, Giovanna Castellano, and Anna M. Fanelli. Meta-data: 
 Characterization of input features for meta-learning. In 2nd International 
 Conference on Modeling Decisions for Artificial Intelligence (MDAI), 
 pages 457 - 468, 2005.

 Shawkat Ali, and Kate A. Smith. On learning algorithm selection for 
 classification. Applied Soft Computing, volume 6, pages 119 - 138, 2006.
}
\seealso{
Other meta-features: 
\code{\link{general}()},
\code{\link{infotheo}()},
\code{\link{itemset}()}
}
\concept{meta-features}
